washer toggle attempt
The idea if it works is that this will create a switch that monitors the status of the machine.  OFF is unit is off, -and- ON is when its in RPP delay status.  When it's ON it can be selected to start the cycle.  Very basic attempt here that needs tested to refine.