Native k-cup support, washer/dryer remote start support, and silence stringprep warning  (#385)
* Silence  stringprep warning

On systems without cython installed the underlying slixmpp library logs a frequent performance warning about stringprep.

This warning is not actionable for most users and creates unnecessary noise in the logs. This change silences the warning by setting the specific logger's level to ERROR, preventing the message from appearing without affecting functionality.

* Add native switch for refrigerator K-Cup hot water

This adds a new switch entity to control the K-Cup/hot water brewing feature on compatible GE fridges.

This change makes it a native switch that is automatically discovered, providing a user-friendly experience directly on the device card.

The new switch is only created for devices that report having the hot water feature, so it shouldn't affect users with other appliance models.

* washer toggle attempt

The idea if it works is that this will create a switch that monitors the status of the machine.  OFF is unit is off, -and- ON is when its in RPP delay status.  When it's ON it can be selected to start the cycle.  Very basic attempt here that needs tested to refine.

* Update __init__.py

Forgot to push the line.

* Full implementation of RPP control switches

This adds RPP control switches to washers and dryers with remote delay functionality.  Device must be in the remote delay mode for the button to be available -and- for a press event to work.

* dryer.py update

Adding to have the button exactly as the washer.

* Additional model support

Some models do not report "Delay Run" as a status code.  Changes made that permit support for more models.